Update task desrtoy view. Approved task cannot be deleted.

+++ task/views/general_task_views.py
@@ -33,8 +33,10 @@ class TaskViewSet(ViewSet, PageNumberPagination):
     def destroy(self, request, pk):
         task = self.get_object(pk)
         self.check_object_permissions(request, task)
-        task.delete()
-        return Response(data={"detail": [_("Task delete successful")]}, status=status.HTTP_202_ACCEPTED)
+        if task.approval_status == Task.ApprovalChoices.APPROVED:
+            task.delete()
+            return Response(data={"detail": [_("Task delete successful")]}, status=status.HTTP_202_ACCEPTED)
+        return Response(data={"detail": [_("Task canont be deleted after approval from department head")]})
 
     def retrieve(self, request, pk):
         task = self.get_object(pk)
